Add --format json to cli paper search
